fix rollup, further restructure tests
