+++ sdk/communication/communication-sms/test/internal/smsClient.internal.mocked.spec.ts
