+++ sdk/communication/communication-sms/test/internal/smsClient.internal.spec.ts
