+++ sdk/communication/communication-sms/test/public/node/smsClientWithToken.spec.ts
@@ -1,41 +1,26 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-/**
- *  These are a duplicate of the public SmsClientWithToken.spec.ts tests, but with additional logic to enable recording/playback.
- *  This is a workaround because Http Requests with Randomized UUIDs do not play well with the recorder
- *  These tests will be skipped in Live Mode since the public tests run in live mode only.
- */
-
-import { env, isLiveMode, isPlaybackMode, record, Recorder } from "@azure/test-utils-recorder";
-import { SmsSendRequest, SmsClient } from "../../src/smsClient";
+import { env, isPlaybackMode, record, Recorder } from "@azure/test-utils-recorder";
+import { SmsSendRequest, SmsClient } from "../../../src/smsClient";
 import { assert } from "chai";
 import { isNode } from "@azure/core-http";
 import * as dotenv from "dotenv";
 import * as sinon from "sinon";
 import { parseConnectionString } from "@azure/communication-common";
-import { createCredential, recorderConfiguration } from "../utils/recordedClient";
-import { Uuid } from "../../src/utils/uuid";
+import { createCredential, recorderConfiguration } from "../../utils/recordedClient";
 import { Context } from "mocha";
 
 if (isNode) {
   dotenv.config();
 }
 
-describe("SmsClientWithToken [Playback/Record]", async () => {
+describe("SmsClientWithToken [Live]", async () => {
   let recorder: Recorder;
 
   beforeEach(async function(this: Context) {
-    if (isLiveMode()) {
-      console.log("Skipping because public tests will run instead.");
-      this.skip();
-    }
     recorder = record(this, recorderConfiguration);
-    recorder.skip("browser");
-    if (isPlaybackMode()) {
-      sinon.stub(Uuid, "generateUuid").returns("sanitized");
-      sinon.stub(Date, "now").returns(0);
-    }
+    recorder.skip(undefined); //only run in live mode
   });
 
   afterEach(async function(this: Context) {
